Extract currency create endpoint into dedicated controller
